+++ src/Frame.php
@@ -13,6 +13,9 @@ final class Frame
 {
     public const INTERNAL_FRAME_FILENAME = '[internal]';
 
+    /**
+     * @deprecated This constant is deprecated and will be removed in 5.x.
+     */
     public const ANONYMOUS_CLASS_PREFIX = "class@anonymous\x00";
 
     /**