+++ src/Tracing/DynamicSamplingContext.php
@@ -189,6 +189,10 @@ final class DynamicSamplingContext
             $samplingContext->set('sampled', $transaction->getSampled() ? 'true' : 'false');
         }
 
+        if ($transaction->getMetadata()->getSampleRand() !== null) {
+            $samplingContext->set('sample_rand', (string) $transaction->getMetadata()->getSampleRand());
+        }
+
         $samplingContext->freeze();
 
         return $samplingContext;
@@ -198,6 +202,7 @@ final class DynamicSamplingContext
     {
         $samplingContext = new self();
         $samplingContext->set('trace_id', (string) $scope->getPropagationContext()->getTraceId());
+        $samplingContext->set('sample_rand', (string) $scope->getPropagationContext()->getSampleRand());
 
         if ($options->getTracesSampleRate() !== null) {
             $samplingContext->set('sample_rate', (string) $options->getTracesSampleRate());