+++ src/Tracing/TransactionMetadata.php
@@ -22,20 +22,36 @@ final class TransactionMetadata
     private $source;
 
     /**
+     * @var float|int|null
+     */
+    private $parentSamplingRate;
+
+    /**
+     * @var float|int|null
+     */
+    private $sampleRand;
+
+    /**
      * Constructor.
      *
      * @param float|int|null              $samplingRate           The sampling rate
      * @param DynamicSamplingContext|null $dynamicSamplingContext The Dynamic Sampling Context
      * @param TransactionSource|null      $source                 The transaction source
+     * @param float|null                  $parentSamplingRate     The parent sampling rate
+     * @param float|null                  $sampleRand             The trace sample rand
      */
     public function __construct(
         $samplingRate = null,
         ?DynamicSamplingContext $dynamicSamplingContext = null,
-        ?TransactionSource $source = null
+        ?TransactionSource $source = null,
+        ?float $parentSamplingRate = null,
+        ?float $sampleRand = null
     ) {
         $this->samplingRate = $samplingRate;
         $this->dynamicSamplingContext = $dynamicSamplingContext;
         $this->source = $source ?? TransactionSource::custom();
+        $this->parentSamplingRate = $parentSamplingRate;
+        $this->sampleRand = $sampleRand ?? round(mt_rand(0, mt_getrandmax() - 1) / mt_getrandmax(), 6);
     }
 
     /**
@@ -52,6 +68,30 @@ final class TransactionMetadata
     public function setSamplingRate($samplingRate): self
     {
         $this->samplingRate = $samplingRate;
+
+        return $this;
+    }
+
+    public function getParentSamplingRate(): ?float
+    {
+        return $this->parentSamplingRate;
+    }
+
+    public function setParentSamplingRate(?float $parentSamplingRate): self
+    {
+        $this->parentSamplingRate = $parentSamplingRate;
+
+        return $this;
+    }
+
+    public function getSampleRand(): ?float
+    {
+        return $this->sampleRand;
+    }
+
+    public function setSampleRand(?float $sampleRand): self
+    {
+        $this->sampleRand = $sampleRand;
 
         return $this;
     }